Add --fix-mismatch, improve --download
Added the option --fix-mismatch which reads the latest logfile and checks the version of the servers files against the clients files. If files are missing these are downloaded, if files are mismatched the old file is removed and the new file is downloaded. If no redirect is specified nothing is done.

Some small improvements to download: files are only decompressed if these are .uz files and a case insensitive file search is added