0.4 Multidir and improved Cache
Added support for finding files in multiple directories, the first directory in @utdirs is where all the files are stored, the directory in which System/uz-get.ini is and in which ucc dumps its files. On linux this is ~/.loki/ut/

Added Caching of mismatched files, these will be stored in $utdirs[0]/Cache/ with the guid appended to the name. On finding mismatched packages uz-get will look into the Cache first.